reassign tasks into where they belong
