+++ src/bin/event_orderbook.rs
@@ -1,7 +1,8 @@
 /// Syncs orderbook
 use kucoin_api::client::{Kucoin, KucoinEnv};
 use kucoin_api::model::websocket::WSTopic;
-use kucoin_arbitrage::broker::orderbook::kucoin::{task_pub_orderbook_event, task_sync_orderbook};
+use kucoin_arbitrage::broker::orderbook::kucoin::task_pub_orderbook_event;
+use kucoin_arbitrage::broker::orderbook::internal::task_sync_orderbook;
 use kucoin_arbitrage::model::counter::Counter;
 use kucoin_arbitrage::model::orderbook::FullOrderbook;
 use std::sync::Arc;
