+++ src/broker/orderbook/kucoin.rs
@@ -1,18 +1,18 @@
 use crate::event::orderbook::OrderbookEvent;
-use crate::global::counter_helper;
-use crate::model::counter::Counter;
 use crate::model::orderbook::FullOrderbook;
-use crate::translator::traits::OrderBookChangeTranslator;
+use crate::model::symbol::SymbolInfo;
+use crate::translator::traits::{OrderBookChangeTranslator, OrderBookTranslator};
 use kucoin_api::client::Kucoin;
 use kucoin_api::futures::TryStreamExt;
+use kucoin_api::model::market::{OrderBook, OrderBookType};
 use kucoin_api::model::websocket::{KucoinWebsocketMsg, WSTopic, WSType};
 use std::sync::Arc;
-use tokio::sync::broadcast::{Receiver, Sender};
+use tokio::sync::broadcast::Sender;
 use tokio::sync::Mutex;
+use tokio::task::JoinSet;
+use tokio::time::Duration;
 
-/// Task to publish orderbook events.
-/// Subscribes Websocket API.
-/// Publishes OrderbookEvent directly after conversion.
+/// Subscribe Websocket API, then publish OrderbookEvent directly after conversion to internal model.
 pub async fn task_pub_orderbook_event(
     api: Kucoin,
     topics: Vec<WSTopic>,
@@ -51,46 +51,66 @@ pub async fn task_pub_orderbook_event(
     }
 }
 
-/// Task to sync local orderbook from API.
-/// Subscribes OrderbookEvent.
-/// Publishes OrderbookEvent after sync.
-pub async fn task_sync_orderbook(
-    mut receiver: Receiver<OrderbookEvent>,
-    sender: Sender<OrderbookEvent>,
-    local_full_orderbook: Arc<Mutex<FullOrderbook>>,
-    counter: Arc<Mutex<Counter>>,
-) -> Result<(), failure::Error> {
+/// Obtain current orderbook of a list of symbol from Kucoin REST API
+pub async fn task_get_orderbook(api: Kucoin, symbol: &str) -> Result<OrderBook, failure::Error> {
+    let mut try_counter = 0;
     loop {
-        counter_helper::increment(counter.clone()).await;
-        let event = receiver.recv().await?;
-        let mut full_orderbook = local_full_orderbook.lock().await;
-        match event {
-            OrderbookEvent::OrderbookReceived((symbol, orderbook)) => {
-                (*full_orderbook).insert(symbol.clone(), orderbook);
-                log::info!("Initialised Orderbook for {symbol}")
+        try_counter += 1;
+        // OrderBookType::Full requires valid API Key
+        let res = api.get_orderbook(symbol, OrderBookType::L20).await;
+        if res.is_err() {
+            log::warn!("orderbook[{symbol}] did not respond ({try_counter:?} tries)");
+            continue;
+        }
+        let res = res.unwrap();
+        let code = res.code;
+        match code.as_str() {
+            "429000" => {
+                log::warn!("[{symbol:?}] request overloaded ({try_counter:?} tries)")
             }
-            OrderbookEvent::OrderbookChangeReceived((symbol, orderbook_change)) => {
-                let orderbook = (*full_orderbook).get_mut(&symbol);
-                if orderbook.is_none() {
-                    return Err(failure::err_msg(format!(
-                        "received {symbol} but orderbook not initialised yet."
-                    )));
+            "200000" => {
+                if res.data.is_none() {
+                    log::warn!("orderbook[{symbol}] received none ({try_counter:?} tries)");
+                    continue;
                 }
-                // log::info!("insertion: {orderbook_change:#?}");
-                match orderbook.unwrap().merge(orderbook_change) {
-                    Ok(res) => {
-                        if let Some(ob) = res {
-                            // log::info!("update: {ob:#?}");
-                            sender
-                                .send(OrderbookEvent::OrderbookChangeReceived((symbol, ob)))
-                                .unwrap();
-                        }
-                    }
-                    Err(e) => {
-                        log::error!("Merge conflict: {e}")
-                    }
-                }
+                log::info!("obtained [{symbol}]");
+                return Ok(res.data.unwrap());
             }
+            "400003" => return Err(failure::err_msg("API key needed not but provided")),
+            _ => return Err(failure::err_msg(format!("unrecognised code [{:?}]", code))),
         }
     }
+}
+
+/// Obtain all the inital orderbook using Kucoin REST API
+pub async fn task_get_initial_orderbooks(
+    api: Kucoin,
+    symbol_infos: Vec<SymbolInfo>,
+    full_orderbook: Arc<Mutex<FullOrderbook>>,
+) -> Result<(), failure::Error> {
+    // replace spawn with or a taskpool
+    let mut taskpool_aggregate = JoinSet::new();
+    // collect all initial orderbook states with REST
+    let symbols: Vec<String> = symbol_infos.into_iter().map(|info| info.symbol).collect();
+    log::info!("Total symbols: {:?}", symbols.len());
+    for symbol in symbols {
+        let api = api.clone();
+        let full_orderbook_arc = full_orderbook.clone();
+        // further improve performance and the server overload issue with mixed use of select and jjoin
+        taskpool_aggregate.spawn(async move {
+            let data = task_get_orderbook(api, &symbol).await.unwrap();
+            let mut x = full_orderbook_arc.lock().await;
+            x.insert(symbol.to_string(), data.to_internal());
+            symbol
+        });
+        // prevent server overloading
+        tokio::time::sleep(Duration::from_millis(25)).await;
+    }
+    while let Some(res) = taskpool_aggregate.join_next().await {
+        if let Err(e) = res {
+            return Err(failure::Error::from(e));
+        }
+        log::info!("Initialized orderbook for [{:?}]", res.unwrap());
+    }
+    Ok(())
 }
