+++ src/broker/orderbook/mod.rs
@@ -1,1 +1,4 @@
+/// Obtaining orderbook state and changes using Kucoin API
 pub mod kucoin;
+/// Syncing with local copies of orderbook.
+pub mod internal;
